Move out continuation utilities
